requirements.txt
aiogram==2.25.1
alembic==1.13.0
APScheduler==3.10.4
pandas==2.1.4
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
